#20 Rename plant subpackage to stage
