+++ stage/doc.go
@@ -1,5 +1,5 @@
-// Package plant provides description of data flow, which takes a kind of
+// Package stage provides description of data flow, which takes a kind of
 // transformation of its values. The description separates the transformation
 // (_forming_ in terms of the package) from their transporting, and allows to
 // create data flows with help of the transformers' implementations.
-package plant
+package stage
