+++ stage/new.go
@@ -1,4 +1,4 @@
-package plant
+package stage
 
 import (
 	"context"
